rename: domain 네이밍 오타 변경
