+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
