+++ src/main/java/blackjack/domain/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.Arrays;
 import java.util.Collections;
