+++ src/main/java/blackjack/domain/Players.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.List;
 import java.util.function.Consumer;
