+++ src/main/java/blackjack/domain/RandomCardsShuffler.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.Collections;
 import java.util.List;
