+++ src/main/java/blackjack/domain/Rank.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.List;
