+++ src/main/java/blackjack/domain/ScoreCalculator.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.List;
