+++ src/main/java/blackjack/domain/Suit.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 public enum Suit {
     SPADE, HEART, DIAMOND, CLUB
