+++ src/main/java/blackjack/domain/WinnerDecider.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 import java.util.List;
 import java.util.Set;
