+++ src/main/java/blackjack/domain/WinningResult.java
@@ -1,4 +1,4 @@
-package blackjack.domian;
+package blackjack.domain;
 
 public enum WinningResult {
     DRAW, WIN, LOSE
