+++ src/test/java/blackjack/domain/CardTypes.java
@@ -1,8 +1,5 @@
 package blackjack.domain;
 
-import blackjack.domian.Card;
-import blackjack.domian.Rank;
-import blackjack.domian.Suit;
 import java.util.ArrayList;
 import java.util.List;
