+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,9 +1,5 @@
 package blackjack.domain;
 
-import blackjack.domian.Card;
-import blackjack.domian.Deck;
-import blackjack.domian.Rank;
-import blackjack.domian.Suit;
 import java.util.List;
 import java.util.Stack;
 import org.assertj.core.api.Assertions;
