+++ src/test/java/blackjack/domain/FixCardsShuffler.java
@@ -1,12 +1,10 @@
 package blackjack.domain;
 
-import blackjack.domian.Card;
-import blackjack.domian.CardsShuffler;
 import java.util.List;
 
 public class FixCardsShuffler implements CardsShuffler {
     @Override
     public void shuffle(List<Card> cards) {
-        
+
     }
 }
